backend/requirements.txt
# Core Django
Django>=5.1,<5.2
django-ninja>=1.3.0
psycopg2-binary>=2.9.9

# Authentication & Security
django-cors-headers>=4.3.1
PyJWT>=2.8.0
python-dotenv>=1.0.0

# Email
django-anymail[amazon-ses]>=11.0

# Production server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Environment & Config
dj-database-url>=2.1.0

# Utilities
python-dateutil>=2.8.2
